3 Class updates, 1 new Class
I wrote a HitOrMiss class that changes the tile on the enemy board accordingly and also returns a hitToken (I'm not sure if we still need it but I left it in). I also updated PlayerActions to include the attack(), heal(), and I also threw hitOrMiss in there as well just in case we decide we like that better. I also played around with the GameMain class that we discussed in meeting today, so I extracted some of the code from BattleshipBoard and it is now in GameMain. They seem to be working fine, other than that the player1 board and player2 board are acting as the same object. I think it will be a simple fix, but I'm going to work on something else now.