Update BattleshipBoard.java
made a whoopsie because at the end of the meeting we thought it'd be cute if we used ~ instead off . for empty spaces. It was cute, but it broke the code because we didnt go through and change all the checks. Im working on the code right now and just realised what we had done. I've set it back to a . for now, once I figure out this overlapping ship problem Ill go back to the cute ~